Update authorization middleware, add in title/assignees/labels to issue templates (based off #706)
