+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,9 +1,9 @@
 ---
 name: Bug report
 about: Create a report to help us improve
-title: ''
+title: '[Bug Report] '
 labels: bug
-assignees: ''
+assignees: 'auguwu'
 
 ---
