+++ src/commands/core/StatisticsCommand.ts
@@ -35,61 +35,6 @@ import Config from '../../components/Config';
 import Redis from '../../components/Redis';
 import os from 'os';
 
-interface RedisInfo {
-  total_connections_received: number;
-  total_commands_processed: number;
-  instantaneous_ops_per_sec: number;
-  total_net_input_bytes: number;
-  total_net_output_bytes: number;
-  instantaneous_input_kbps: number;
-  instantaneous_output_kbps: number;
-  rejected_connections: number;
-  sync_full: number;
-  sync_partial_ok: number;
-  sync_partial_err: number;
-  expired_keys: number;
-  expired_stale_perc: number;
-  expired_time_cap_reached_count: number;
-  evicted_keys: number;
-  keyspace_hits: number;
-  keyspace_misses: number;
-  pubsub_channels: number;
-  pubsub_patterns: number;
-  latest_fork_usec: number;
-  migrate_cached_sockets: number;
-  slave_expires_tracked_keys: number;
-  active_defrag_hits: number;
-  active_defrag_misses: number;
-  active_defrag_key_hits: number;
-  active_defrag_key_misses: number;
-}
-
-interface RedisServerInfo {
-  redis_version: string;
-  redis_git_sha1: string;
-  redis_git_dirty: string;
-  redis_build_id: string;
-  redis_mode: string;
-  os: string;
-  arch_bits: string;
-  multiplexing_api: string;
-  atomicvar_api: string;
-  gcc_version: string;
-  process_id: string;
-  process_supervised: string;
-  run_id: string;
-  tcp_port: string;
-  server_time_usec: string;
-  uptime_in_seconds: string;
-  uptime_in_days: string;
-  hz: string;
-  configured_hz: string;
-  lru_clock: string;
-  executable: string;
-  config_file: string;
-  io_threads_active: string;
-}
-
 export default class StatisticsCommand extends Command {
   private parent!: CommandService;
 
@@ -114,76 +59,9 @@ export default class StatisticsCommand extends Command {
     });
   }
 
-  private async getRedisInfo() {
-    const stopwatch = new Stopwatch();
-    stopwatch.start();
-    await this.redis.client.ping('Ice is cute as FUCK');
-
-    const ping = stopwatch.end();
-
-    // stole this from donny
-    // Credit: https://github.com/FurryBotCo/FurryBot/blob/master/src/commands/information/stats-cmd.ts#L22
-    const [stats, server] = await Promise.all([
-      this.redis.client.info('stats').then(info =>
-        info
-          .split(/\n\r?/)
-          .slice(1, -1)
-          .map(item => ({ [item.split(':')[0]]: item.split(':')[1].trim() }))
-          .reduce((a, b) => ({ ...a, ...b })) as unknown as RedisInfo
-      ),
-
-      this.redis.client.info('server').then(info =>
-        info
-          .split(/\n\r?/)
-          .slice(1, -1)
-          .map(item => ({ [item.split(':')[0]]: item.split(':')[1].trim() }))
-          .reduce((a, b) => ({ ...a, ...b })) as unknown as RedisServerInfo
-      )
-    ]);
-
-    return {
-      server,
-      stats,
-      ping
-    };
-  }
-
-  async getDatabaseStatistics() {
-    const stopwatch = new Stopwatch();
-    stopwatch.start();
-    await this.database.connection.query('SELECT * FROM guilds');
-    const ping = stopwatch.end();
-
-    let dbName: string = 'nino';
-    const url = this.config.getProperty('database.url');
-    if (url !== undefined) {
-      const parts = url.split('/');
-      dbName = parts[parts.length - 1];
-    } else {
-      dbName = this.config.getProperty('database.database') ?? 'nino';
-    }
-
-    // collect shit
-    const data = await Promise.all([
-      this.database.connection.query(`SELECT tup_returned, tup_fetched, tup_inserted, tup_updated, tup_deleted FROM pg_stat_database WHERE datname = '${dbName}';`),
-      this.database.connection.query('SELECT version();'),
-      this.database.connection.query('SELECT extract(epoch FROM current_timestamp - pg_postmaster_start_time()) AS uptime;')
-    ]);
-
-    return {
-      inserted: Number(data[0]?.[0]?.tup_inserted ?? 0),
-      updated: Number(data[0]?.[0]?.tup_updated ?? 0),
-      deleted: Number(data[0]?.[0]?.tup_deleted ?? 0),
-      fetched: Number(data[0]?.[0]?.tup_fetched ?? 0),
-      version: data[1][0].version.split(', ').shift().replace('PostgreSQL ', '').trim(),
-      uptime: humanize(Math.floor(data[2][0].uptime * 1000), true),
-      ping
-    };
-  }
-
   async run(msg: CommandMessage) {
-    const database = await this.getDatabaseStatistics();
-    const redis = await this.getRedisInfo();
+    const database = await this.database.getStatistics();
+    const redis = await this.redis.getStatistics();
     const guilds = this.discord.client.guilds.size.toLocaleString();
     const users = this.discord.client.guilds.reduce((a, b) => a + b.memberCount, 0).toLocaleString();
     const channels = Object.keys(this.discord.client.channelGuildMap).length.toLocaleString();
